% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCME guideline calculations.R
\name{calc_water_guidelines}
\alias{calc_water_guidelines}
\title{Calculate CCME water quality guidelines for compounds with Equations}
\usage{
calc_water_guidelines(
  Parameter = c("Manganese", "Zinc", "Cadmium", "Nickel", "Copper", "Lead"),
  Hardness = NA,
  pH = NA,
  DOC = NA
)
}
\arguments{
\item{Parameter}{Options include "Manganese", "Zinc", "Cadmium", "Nickel", "Copper", "Lead"}

\item{Hardness}{Required for all (mg/L)}

\item{pH}{Required for Manganese}

\item{DOC}{Required for Zinc (mg/L)}
}
\value{
Returns a vectors with the short-term, long-term limits and whether the limit if for Dissolved or Total
}
\description{
Calculate CCME water quality guidelines for compounds with Equations
}
